requirements.txt
Flask==2.3.2
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
keras==2.12.0
tensorflow==2.12.0
Werkzeug==2.3.6 